Add Kalshi and Polymarket venue adapters
Both adapters reduce their venue to one interface so strategies, risk and the
backtester are written once. API contracts were verified against the official
SDK sources (kalshi-python 2.1.0, py-clob-client) rather than prose docs, since
the generated clients are the more reliable spec.

The load-bearing piece is book normalisation. Kalshi publishes bids on *both*
sides of a binary market and no asks at all, so YES asks must be derived by
reflecting NO bids through $1. Getting that backwards raises no error - it
silently inverts the book and every strategy trades the wrong way. It is
isolated in one function with property tests, including one asserting that a
crossed book appears exactly when yes_bid + no_bid >= $1, which is the real
YES/NO parity arbitrage condition rather than a data error.

Kalshi specifics: RSA-PSS request signing with salt_length=DIGEST_LENGTH over
timestamp+method+path. Because the signature covers a timestamp, clock drift
presents as blanket 401s that look like bad credentials, so clock skew is
measured explicitly against the server Date header and trading refuses to start
beyond 5s. Short-position cost basis is recovered from Kalshi's collateral
figure as exposure - $1 * size.

Polymarket splits read and write paths deliberately. Public book/price
endpoints go through our own async layer for the hot path; order placement
delegates to the official client because it requires an EIP-712 signature over
the CTF Exchange order struct, where reimplementation risks silently malformed
signatures for no benefit. Notably Polymarket supports post_only and
split/merge, which Kalshi does not - both matter for strategy selection.

67 tests passing.

+++ src/orbit/core/money.py
@@ -68,7 +68,7 @@ def dollars_to_pips(dollars: float) -> int:
 
     Used to ingest venue payloads that quote decimal dollars (Polymarket).
     """
-    return int(round(dollars * PIPS_PER_DOLLAR))
+    return round(dollars * PIPS_PER_DOLLAR)
 
 
 def pips_to_dollars(pips: int) -> float:
